bug-fix:

+++ src/main/java/com/amanov/payrollapp/controllers/PayrollController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @RestController
@@ -18,7 +19,7 @@ public class PayrollController {
 
     @GetMapping("/calculate")
     public ResponseEntity<PaymentResponse> getPayroll(
-            @RequestParam(name = "salary") Double salary,
+            @RequestParam(name = "salary") BigDecimal salary,
             @RequestParam(name = "firstDay") LocalDate firstDay,
             @RequestParam(name = "lastDay") LocalDate lastDay) {
         PaymentRequest paymentRequest = new PaymentRequest(salary, firstDay, lastDay);
