+++ src/main/java/com/amanov/payrollapp/dto/PaymentRequest.java
@@ -3,12 +3,13 @@ package com.amanov.payrollapp.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 @AllArgsConstructor
 public class PaymentRequest {
-    private Double salary;
+    private BigDecimal salary;
     private LocalDate firstDay;
     private LocalDate lastDay;
 }
