+++ src/test/java/com/amanov/payrollapp/PayrollAppApplicationTests.java
@@ -8,6 +8,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @SpringBootTest
@@ -17,7 +18,7 @@ class PayrollAppApplicationTests {
 
     @Test
     void calculate() {
-        Double salary = 100000D;
+        BigDecimal salary = BigDecimal.valueOf(100000);
         LocalDate firstDay = LocalDate.of(2023, 5, 1);
         LocalDate lastDay = LocalDate.of(2023, 5, 15);
         HttpStatusCode statusCode = controller.getPayroll(salary, firstDay, lastDay).getStatusCode();
